[math] remove the hard requirement of `taichi`
