+++ brainpy/_src/math/brainpylib_check.py
@@ -2,25 +2,49 @@ import os
 import platform
 import ctypes
 
-import taichi as ti
 from jax.lib import xla_client
 
-taichi_path = ti.__path__[0]
-taichi_c_api_install_dir = os.path.join(taichi_path, '_lib', 'c_api')
-os.environ['TAICHI_C_API_INSTALL_DIR'] = taichi_c_api_install_dir
-os.environ['TI_LIB_DIR'] = os.path.join(taichi_c_api_install_dir, 'runtime')
 
-# link DLL
-if platform.system() == 'Windows':
-  try:
-    ctypes.CDLL(taichi_c_api_install_dir + '/bin/taichi_c_api.dll')
-  except OSError:
-    raise OSError(f'Does not found {taichi_c_api_install_dir + "/bin/taichi_c_api.dll"}')
-elif platform.system() == 'Linux':
-  try:
-    ctypes.CDLL(taichi_c_api_install_dir + '/lib/libtaichi_c_api.so')
-  except OSError:
-    raise OSError(f'Does not found {taichi_c_api_install_dir + "/lib/taichi_c_api.dll"}')
+try:
+  import taichi as ti
+except (ImportError, ModuleNotFoundError):
+  ti = None
+
+
+def import_taichi():
+  if ti is None:
+    raise ModuleNotFoundError(
+      'Taichi is needed. Please install taichi through:\n\n'
+      '> pip install -i https://pypi.taichi.graphics/simple/ taichi-nightly'
+    )
+  if ti.__version__ < (1, 7, 0):
+    raise RuntimeError(
+      'We need taichi>=1.7.0. Currently you can install taichi>=1.7.0 through taichi-nightly:\n\n'
+      '> pip install -i https://pypi.taichi.graphics/simple/ taichi-nightly'
+    )
+  return ti
+
+
+if ti is None:
+  is_taichi_installed = False
+else:
+  is_taichi_installed = True
+  taichi_path = ti.__path__[0]
+  taichi_c_api_install_dir = os.path.join(taichi_path, '_lib', 'c_api')
+  os.environ['TAICHI_C_API_INSTALL_DIR'] = taichi_c_api_install_dir
+  os.environ['TI_LIB_DIR'] = os.path.join(taichi_c_api_install_dir, 'runtime')
+
+  # link DLL
+  if platform.system() == 'Windows':
+    try:
+      ctypes.CDLL(taichi_c_api_install_dir + '/bin/taichi_c_api.dll')
+    except OSError:
+      raise OSError(f'Can not find {taichi_c_api_install_dir + "/bin/taichi_c_api.dll"}')
+  elif platform.system() == 'Linux':
+    try:
+      ctypes.CDLL(taichi_c_api_install_dir + '/lib/libtaichi_c_api.so')
+    except OSError:
+      raise OSError(f'Can not find {taichi_c_api_install_dir + "/lib/taichi_c_api.dll"}')
 
 # Register the CPU XLA custom calls
 try:
